Implement setLogger
This sets a function to receive logs from the headset. By default all the
logs will be output to the stdout/stderr. However in certain scenarios
users may want to redirect these logs to another stream.

This method gives one the ability to specify a function that will receive
any logs coming from the library, so the user chooses where and how to
display the logs.